# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "description" : "Household task management API.\n\nConventions callers need:\n- Every response is the envelope `{ success, message, data }`.\n- `user`, `group` and `task` are addressed by opaque `uid`; every other resource by numeric `id`.\n- Timestamps are UTC with no zone suffix — parse as UTC, convert for display.\n- Actors are derived from the token; never send a user or member id to identify the caller.\n- Errors: 400 validation, 401 unauthenticated, 403 forbidden, 404 not found, 409 duplicate or violated business rule, 429 rate limited (see `Retry-After`).\n- Role requirements are stated per operation and are not modelled in this spec.\n",
    "title" : "TasKlean API",
    "version" : "0.0.1-SNAPSHOT"
  },
  "servers" : [ {
    "url" : "/",
    "description" : "Relative to the host serving this API"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/task-completions" : {
      "get" : {
        "description" : "Lists a task's completion history, newest first.",
        "operationId" : "getCompletionsByTask",
        "parameters" : [ {
          "description" : "the task id",
          "in" : "query",
          "name" : "taskId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListTaskCompletionResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the completions"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Lists a task's completion history, newest first.",
        "tags" : [ "Task completions" ]
      },
      "post" : {
        "description" : "Records a task completion.",
        "operationId" : "createCompletion",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TaskCompletionRequest"
              }
            }
          },
          "description" : "the completion details (task, member, optional photo/note)",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTaskCompletionResponse"
                }
              }
            },
            "description" : "<code>201 Created</code> with the recorded completion"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Records a task completion.",
        "tags" : [ "Task completions" ]
      }
    },
    "/api/task-completions/member/{groupMemberId}" : {
      "get" : {
        "description" : "Lists the completions recorded by a group member, newest first.",
        "operationId" : "getCompletionsByMember",
        "parameters" : [ {
          "description" : "the group member id",
          "in" : "path",
          "name" : "groupMemberId",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListTaskCompletionResponse"
                }
              }
            },
            "description" : "<code>200 OK</code> with the completions"
          },
          "400" : {
            "$ref" : "#/components/responses/BadRequest"
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "429" : {
            "$ref" : "#/components/responses/TooManyRequests"
          },
          "500" : {
            "$ref" : "#/components/responses/InternalServerError"
          }
        },
        "summary" : "Lists the completions recorded by a group member, newest first.",
        "tags" : [ "Task completions" ]
      }
    },
...
  },
  "components" : {
    "responses" : {
      "BadRequest" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "Validation failure, malformed body, or a bad path/query parameter"
      },
      "Forbidden" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "Authenticated, but lacking the platform or group role this operation requires"
      },
      "InternalServerError" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "Unexpected server error"
      },
      "TooManyRequests" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "Rate limit exceeded — see the Retry-After header"
      },
      "Unauthorized" : {
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ErrorResponse"
            }
          }
        },
        "description" : "No valid access token — or, on the public auth endpoints, rejected credentials"
      }
    },
    "schemas" : {
...
      "ApiResponseListTaskCompletionResponse" : {
        "type" : "object",
        "description" : "Standard response envelope wrapping every API result with a <code>success</code> flag,\n an optional <code>message</code>, and the payload <code>data</code>. Construct via the static\n factory methods rather than the builder directly.",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TaskCompletionResponse"
            }
          },
          "message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
...
      "ApiResponseTaskCompletionResponse" : {
        "type" : "object",
        "description" : "Standard response envelope wrapping every API result with a <code>success</code> flag,\n an optional <code>message</code>, and the payload <code>data</code>. Construct via the static\n factory methods rather than the builder directly.",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/TaskCompletionResponse"
          },
          "message" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        }
      },
...
      "ErrorResponse" : {
        "type" : "object",
        "description" : "The standard envelope, as returned for a failed request.",
        "properties" : {
          "data" : {
            "type" : "object",
            "description" : "Always null on an error"
          },
          "message" : {
            "type" : "string",
            "description" : "Human-readable error message"
          },
          "success" : {
            "type" : "boolean",
            "default" : false
          }
        }
      },
...
      "TaskCompletionRequest" : {
        "type" : "object",
        "properties" : {
          "completionNote" : {
            "type" : "string"
          },
          "completionPhotoUrl" : {
            "type" : "string"
          },
          "groupMemberId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "taskId" : {
            "type" : "integer",
            "format" : "int64"
          }
        },
        "required" : [ "groupMemberId", "taskId" ]
      },
      "TaskCompletionResponse" : {
        "type" : "object",
        "properties" : {
          "completionNote" : {
            "type" : "string"
          },
          "completionPhotoUrl" : {
            "type" : "string"
          },
          "dateCompleted" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dateCreated" : {
            "type" : "string",
            "format" : "date-time"
          },
          "groupMemberId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "taskId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from /api/auth/login, /api/auth/verify-email, /api/auth/google or /api/auth/refresh.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  }
}